Add files via upload
This is a very project which is made out all CSS-Bootstrap components like card, carousel, examples etc.
The project depicts a home page of a shipping company. The webpage has come out very professional and modern.
It was fun doing this project and I got to learn many things from this project.